// src/types/summary-type.ts
export type Summary = {
	today: {
		date: string
		workedMinutes: number
		extraMinutes: number
		tips: number
	}
	thisWeek: {
		date: string
		workedMinutes: number
		extraMinutes: number
		tips: number
	}
	thisMonth: {
		date: string
		workedMinutes: number
		extraMinutes: number
		tips: number
	}
}